requirements.txt
flask
openai
ollama
python-dotenv
pymongo
